+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 ## Unreleased
 
+## 1.0.0
+
 - **New**
   - `Token`
     - **Rotate to Target**
     - **Dodge Target**
     - **Shake Token**
     - **Attack Bounce off Token**
+  - `Text`
+    - `Finishing Move - BBB` - Finishing move text from Kekkai Sensen
 
 ## 0.7.3